+++ CHANGELOG.md
@@ -11,9 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## 1.24.0
 ### Added
-- Added 
-  [`output_default`](https://oliver-zehentleitner.github.io/unicorn-binance-websocket-api/unicorn_binance_websocket_api.html?highlight=output_default)
+- [`output_default`](https://oliver-zehentleitner.github.io/unicorn-binance-websocket-api/unicorn_binance_websocket_api.html?highlight=output_default)
   to `BinanceWebSocketApiManager` 
+  [`get_user_agent()`]()
 ## Removed
 - unused import of `ujson` in connection class
 - 4 parameters from `_create_stream_thread`